+++ app/src/components/resources/ToolbarItemNamespaces.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Select, SelectOption, SelectVariant } from '@patternfly/react-core';
 
-import { GetNamespacesRequest, GetNamespacesResponse } from 'proto/clusters_pb';
-import { ClustersPromiseClient } from 'proto/clusters_grpc_web_pb';
+import { ClustersPromiseClient, GetNamespacesRequest, GetNamespacesResponse } from 'proto/clusters_grpc_web_pb';
 import { apiURL } from 'utils/constants';
 
 // clustersService is the Clusters gRPC service, which is used to get all namespaces for the selected clusters.